+++ src/App.tsx
@@ -13,6 +13,7 @@ import { GameLobby } from './pages/GameLobby';
 import { GameSession } from './pages/GameSession';
 import { Leaderboard } from './pages/Leaderboard/Leaderboard';
 import { Login } from './pages/Login/Login';
+import { Logout } from './pages/Logout';
 import { Main } from './pages/Main/Main';
 import { Profile } from './pages/Profile';
 import { ProfileHistory } from './pages/ProfileHistory/ProfileHistory';
@@ -40,6 +41,7 @@ const App: FC<Record<string, never>> = () => {
         <PrivateRoute path="/leaderboard" exact component={Leaderboard} />
         <PrivateRoute path="/profile/history" exact component={ProfileHistory} />
         <PrivateRoute path="/profile" exact component={Profile} />
+        <PrivateRoute path="/logout" exact component={Logout} />
         <Route path="/main" component={Main} />
         <Route path="*" component={Error404} />
       </Switch>
