+++ src/components/Menu.tsx
@@ -1,25 +1,11 @@
 import React, { FC } from 'react';
-import { useHistory } from 'react-router-dom';
 
-import { AuthController } from '../controllers/AuthController';
-import { useAppDispatch } from '../hooks/redux-hooks';
 import { TRANSLATION } from '../lang/ru/translation';
-import { authActions } from '../store/slices/authSlice';
 
-import { Button } from './ui/Button/Button';
 import { MainLink } from './ui/Link/Link';
 import { Title } from './ui/Title/Title';
 
 export const Menu: FC<Record<string, never>> = () => {
-  const dispatch = useAppDispatch();
-  const history = useHistory();
-  const handleLogout = () =>
-    AuthController.logout().then(() => {
-      localStorage.setItem('isAuth', 'false');
-      dispatch(authActions.resetCurrentUser());
-      history.push('/login');
-    });
-
   return (
     <div className="menu">
       <div className="menu__title">
@@ -52,11 +38,10 @@ export const Menu: FC<Record<string, never>> = () => {
             />
           </li>
           <li className="menu__link">
-            <Button
-              extendClass="ui-button--error"
-              type="button"
+            <MainLink
+              extendClass="ui-link--warning"
+              href="/logout"
               text={TRANSLATION.Logout.label}
-              click={handleLogout}
             />
           </li>
         </ul>
